backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
pydantic-settings==2.10.1
python-multipart==0.0.20
PyJWT==2.10.1
